extracted components from Dashboard
