+++ src/components/Footer.css
@@ -1,0 +1,5 @@
+.sticky-footer {
+  padding: 1.5rem;
+  margin: 0 -15px;
+  font-size: 0.75rem;
+}
