+++ src/components/Footer.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './Footer.css';
+
+export default function Footer() {
+  return (
+      <footer className='sticky-footer bg-dark'>
+        <div className='text-center'>
+          <span className='text-muted'>Copyright &copy; 2020 oh CRUD</span>
+        </div>
+      </footer>
+  );
+}
