+++ src/components/Home.css
@@ -1,0 +1,11 @@
+.card {
+  border-left: 5px solid;
+}
+
+.search-box {
+  width: 30%;
+}
+
+/* .react-calendar {
+  min-width: 80%;
+} */
