+++ src/components/NavBar.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-function NavBar() {
+export default function NavBar() {
   return (
-    <nav className='navbar navbar-dark bg-dark fixed-top'>
-      <Link className='navbar-brand' to='/'>
+    <nav className='navbar navbar-dark bg-dark fixed-top flex-md-nowrap'>
+      <Link className='navbar-brand font-weight-bold' to='/'>
         Sneaker Inventory
       </Link>
     </nav>
   );
 }
-
-export default NavBar;
